+++ consensus/spos/commonSubround/export_test.go
@@ -23,10 +23,6 @@ func (sr *SubroundStartRound) InitCurrentRound() bool {
 	return sr.initCurrentRound()
 }
 
-func GetTrimmedPk(pk string) string {
-	return getTrimmedPk(pk)
-}
-
 // subroundBlock
 
 func (sr *SubroundBlock) BlockChain() data.ChainHandler {
