+++ core/constants.go
@@ -1,4 +1,4 @@
 package core
 
-// PkPrefixSize specifies the max numbers of chars to be displayed from one publc key
-const PkPrefixSize = 12
+// pkPrefixSize specifies the max numbers of chars to be displayed from one publc key
+const pkPrefixSize = 12
